优化部分页面并修复已知Bug

+++ GTRead/GTRead/Classes/Analyse/Model/GTReadTimeModel.swift
@@ -15,7 +15,7 @@ struct GTReadTimeModel: Codable {
     }
     
     struct GTReadTimeItem: Codable {
-        var min: Double
+        var min: Int
         
         enum CodingKeys: String, CodingKey {
             case min
